DIABLO-875 Schedule SIS data updates separately from underlying refresh
