+++ diablo/models/development_db.py
@@ -253,8 +253,8 @@ def _set_up_and_run_jobs():
     Job.create(job_schedule_type='day_at', job_schedule_value='16:00', key='canvas')
     Job.create(disabled=True, job_schedule_type='minutes', job_schedule_value='5', key='doomed_to_fail')
     Job.create(is_schedulable=False, job_schedule_type='day_at', job_schedule_value='16:00', key='remind_invitees')
+    Job.create(is_schedulable=False, job_schedule_type='minutes', job_schedule_value='820', key='schedule_updates')
     Job.create(is_schedulable=False, job_schedule_type='minutes', job_schedule_value='720', key='semester_start')
-    Job.create(is_schedulable=False, job_schedule_type='minutes', job_schedule_value='820', key='sis_data_refresh')
     background_job_manager.start(app)
     HouseKeepingJob(app_context=simply_yield).run()
     CanvasJob(app_context=simply_yield).run()
